High Performance C++ Port of Forest Fire Example
The forest-fire.py example runs very slowly on Pi Zero.  Reimplement
the example in C++ for performance.  This provides an example of SPI
bus access to the Unicorn Hat HD from C++ as well as example code for
limiting the rendering FPS.

Tested on a Pi Zero with Unicorn Hat HD.